+++ src/macros/mod.rs
@@ -1,5 +1,13 @@
 mod control;
+mod entry;
+mod env;
 mod fs;
 mod io;
 mod string;
 mod system;
+
+#[cfg(feature = "serde")]
+mod json;
+
+#[cfg(feature = "time")]
+mod time;