+++ src/macros/time.rs
@@ -1,0 +1,13 @@
+#[macro_export]
+macro_rules! now {
+    () => {{
+        $crate::core::now_rfc3339()
+    }};
+}
+
+#[macro_export]
+macro_rules! today {
+    () => {{
+        $crate::core::today_iso()
+    }};
+}